+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.16.10
 * `UploaderComponent::error()` renamed as `getError()`;
 * `Shell::folderChmod()` has a default value for the `$chmod` argument;
+* added `OptionsParser::consume()` method;
 * fixed code and output for the `Shell` class;
 * removed `TestCaseTrait::deleteAllLogs()` method;
 * removed `folderIsWriteable()` global function. Use instead